Verifie l'indentation de gauche ainsi que celle des symboles de variables, fonctions, types et paramètres. Reste structure et union.

+++ test/res/classic.c
@@ -6,8 +6,8 @@ int foo(int argc, int argv)
   return (0);
 }
 
-int bar(int i,
-	int j)
+int            bar(double i,
+		   int j)
 {
   int k;
